Implement #73 for write side
