+++ tests/test_search_without_webserver.rs
@@ -44,13 +44,14 @@ async fn start_server() -> Result<(Arc<WriteSide>, Arc<ReadSide>)> {
                 data_dir: generate_new_path(),
                 embedding_queue_limit: 50,
                 default_embedding_model: OramaModelSerializable(OramaModel::BgeSmall),
+                insert_batch_commit_size: 10_000,
             },
         },
         reader_side: ReadSideConfig {
             input: oramacore::SideChannelType::InMemory,
             config: IndexesConfig {
                 data_dir: generate_new_path(),
-                insert_batch_commit_size: 10,
+                insert_batch_commit_size: 10_000,
             },
         },
     })
